+++ src/app/page.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import CountButton from '@/components/CountButton';
 import StaticButton from '@/components/StaticButton';
 import InheritanceDemo from '@/components/InheritanceDemo';
-import ComponentsInterpolationDemo from '@/components/ComponentsInterpolationDemo';
+import CompositionDemo from '@/components/CompositionDemo';
 import styled from '@/styled.js';
 
 export default function Home() {
@@ -16,7 +16,7 @@ export default function Home() {
       <CountButton />
       <StaticButton />
       <InheritanceDemo />
-      <ComponentsInterpolationDemo />
+      <CompositionDemo />
     </Wrapper>
   );
 }
