+++ src/components/CompositionDemo.js
